refactor: 139 - move filter files into another folder
