+++ src/components/catalog/filter/filter-price/filter-component.ts
@@ -1,9 +1,9 @@
 import { createDropdownFilter } from './filter-dropdown-button';
-import { createPriceFilter } from './filter-price/filter-price';
-import { createDiv, createButton } from '../../utils/create-elements/create-tags';
-import { CATALOG } from '../../pages/catalog/constants';
+import { createPriceFilter } from './filter-price';
+import { createDiv, createButton } from '../../../../utils/create-elements/create-tags';
+import { CATALOG } from '../../../../pages/catalog/constants';
 import { createLengthFilter } from './filter-length';
-import { BUTTONS_CONFIG } from '../../shared/ui-config/button';
+import { BUTTONS_CONFIG } from '../../../../shared/ui-config/button';
 
 export function createFilterComponent(): HTMLDivElement {
   const label = createDiv({ text: 'Filter by:' });
