+++ src/components/catalog/filter/filter-price/style.css
