+++ src/helpers/clear-filters.ts
@@ -1,8 +1,8 @@
 import { queryState } from '../app/state/query-state';
 import { getRangePrices } from '../utils/products-fetch/price-range';
 import { applyQuery } from '../utils/apply-query/apply-query';
-import { clearLengthEmitter } from '../components/catalog/filter-length';
-import { clearPriceEmitter } from '../components/catalog/filter-price/price-range';
+import { clearLengthEmitter } from '../components/catalog/filter/filter-price/filter-length';
+import { clearPriceEmitter } from '../components/catalog/filter/filter-price/price-range';
 
 export async function clearAllFilters(): Promise<void> {
   queryState.filter.length = [];
